+++ data-workflows/handler.py
@@ -2,7 +2,7 @@ import json
 import logging
 
 import activity.processor
-import categories
+import categories.processor
 
 
 LOGGER = logging.getLogger()
@@ -27,5 +27,5 @@ def handle(event, context) -> None:
             version = event.get("version")
             categories_path = event.get("categories_path")
 
-            categories.run_seed_s3_categories_workflow(version, categories_path)
+            categories.processor.seed_s3_categories_workflow(version, categories_path)
             LOGGER.info(f"Update successful for type={event_type}")
